+++ src/main/java/toy/order/domain/member/Member.java
@@ -20,6 +20,6 @@ public class Member {
 
     private String uuid;
 
-    private Double balance;
+    private Integer balance;
 
 }
